requirements.txt
matplotlib<=2.1
numpy<=1.14
Pillow<=5.0
pyserial==3.4
networkx==2.1
scipy==1.1.0

